feat(aws): with aws keys
