+++ manifest_indexing/manifest_indexing_job.py
@@ -5,91 +5,14 @@ Attributes:
 """
 
 import os
-import random
-import string
-import requests
 import json
-import logging
-import boto3
-from datetime import datetime
-from botocore.exceptions import ClientError
 
+from utils import (
+    download_file,
+    upload_file_to_s3_and_generate_presigned_url
+)
 
 from gen3.tools.manifest_indexing import manifest_indexing
-
-
-def randomString(stringLength=10):
-    """Generate a random string of fixed length """
-
-    letters = string.ascii_lowercase
-    print("hello")
-    return "".join(random.choice(letters) for i in range(stringLength))
-
-
-def upload_file(file_name, bucket, object_name=None):
-    """Upload a file to an S3 bucket
-
-    :param file_name: File to upload
-    :param bucket: Bucket to upload to
-    :param object_name: S3 object name. If not specified then file_name is used
-    :return: True if file was uploaded, else False
-    """
-    # If S3 object_name was not specified, use file_name
-    if object_name is None:
-        object_name = file_name
-    # Upload the file
-    s3_client = boto3.client("s3")
-    try:
-        s3_client.upload_file(file_name, bucket, object_name)
-    except ClientError as e:
-        logging.error(e)
-        return False
-    return True
-
-
-def create_presigned_url(bucket_name, object_name, expiration=3600):
-    """Generate a presigned URL to share an S3 object
-
-    :param bucket_name: string
-    :param object_name: string
-    :param expiration: Time in seconds for the presigned URL to remain valid
-    :return: Presigned URL as string. If error, returns None.
-    """
-    # Generate a presigned URL for the S3 object
-    s3_client = boto3.client("s3")
-    try:
-        response = s3_client.generate_presigned_url(
-            "get_object",
-            Params={"Bucket": bucket_name, "Key": object_name},
-            ExpiresIn=expiration,
-        )
-    except ClientError as e:
-        print(e)
-        return None
-    # The response contains the presigned URL
-    return response
-
-
-def _download_file(url, filename):
-    r = requests.get(url)
-    with open(filename, "wb") as f:
-        f.write(r.content)
-
-
-def _upload_and_generate_presigned_url(bucket_name, filename, expiration=3600):
-    """ Upload and generate presigned url"""
-
-    now = datetime.now()
-
-    if log_file and bucket_name:
-        current_time = now.strftime("%m_%d_%y_%H:%M:%S")
-        upload_file_key = "{}_{}.log".format(current_time, randomString(6))
-
-        if upload_file(filename, bucket_name, upload_file_key):
-            presigned_url = create_presigned_url(bucket_name, upload_file_key)
-            return presigned_url
-
-    return None
 
 
 if __name__ == "__main__":
@@ -105,9 +28,11 @@ if __name__ == "__main__":
         indexing_creds.get("indexd_user", "gdcapi"),
         indexing_creds["indexd_password"],
     )
+    aws_access_key_id = indexing_creds.get("aws_access_key_id")
+    aws_secret_access_key = indexing_creds.get("aws_secret_access_key")
 
     filepath = "./manifest_tmp.tsv"
-    _download_file(input_data_json["URL"], filepath)
+    download_file(input_data_json["URL"], filepath)
 
     print("Start to index the manifest ...")
 
@@ -125,14 +50,22 @@ if __name__ == "__main__":
     )
 
     log_file_presigned_url = (
-        _upload_and_generate_presigned_url(indexing_creds.get("bucket"), log_file)
+        upload_file_to_s3_and_generate_presigned_url(
+            indexing_creds["bucket"],
+            log_file,
+            aws_access_key_id,
+            aws_secret_access_key,
+        )
         if log_file
         else None
     )
 
     output_manifest_presigned_url = (
-        _upload_and_generate_presigned_url(
-            indexing_creds.get("bucket"), output_manifest
+        upload_file_to_s3_and_generate_presigned_url(
+            indexing_creds["bucket"],
+            output_manifest,
+            aws_access_key_id,
+            aws_secret_access_key,
         )
         if output_manifest
         else None
